// server/src/validators/sampleValidators.d.ts
export declare const createSampleSchema: z.ZodObject<{
    weight: z.ZodPreprocess<z.ZodNumber>;
    day: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    shift: z.ZodOptional<z.ZodEnum<{
        M: "M";
        E: "E";
    }>>;
    mbj20_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    mbj20_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    jm103_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    tsb: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f1_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f2_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f3_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f4_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f5_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f6_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f7_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f8_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f9_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f10_d4_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f1_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f2_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f3_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f4_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f5_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f6_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f7_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f8_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f9_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f10_d4_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f1_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f2_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f3_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f4_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f5_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f6_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f7_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f8_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f9_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f10_d6_f: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f1_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f2_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f3_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f4_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f5_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f6_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f7_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f8_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f9_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    f10_d6_s: z.ZodPreprocess<z.ZodOptional<z.ZodNumber>>;
    remarks: z.ZodOptional<z.ZodString>;
}, z.core.$strip>;